requirements.txt
numpy==1.23.5

--find-links https://download.pytorch.org/whl/cu117
torch==1.13.0
torchvision==0.13.0

Pillow==9.3.0
matplotlib==3.6.2
absl-py==1.3.0
imageio==2.25.0
pandas==1.5.3
scikit-image==0.22.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
scipy==1.10.1
seaborn==0.13.0